+++ jest.config.js
@@ -3,4 +3,7 @@
 module.exports = {
   roots: ['<rootDir>'],
   testMatch: ['<rootDir>/test/**/*.{spec,test}.{js,jsx,ts,tsx}'],
+  // testMatch: ['<rootDir>/test/ctx/'],
+  // testMatch: ['<rootDir>/test/api/'],
+  // testMatch: ['<rootDir>/test/api-with-comp/'],
 };