+++ src/support/util.js
@@ -530,7 +530,7 @@ export function isSymbol(value) {
   )
 }
 
-export function delay(ms = 0) {
+export function delay(ms = 1000) {
   return new Promise(resolve => setTimeout(resolve, ms));
 }