+++ test/ctx/_empty.test.js
@@ -1,5 +1,7 @@
 /** for copy */
+/** @typedef {import('../../src/types-inner').IRefCtx} Ctx */
 import React from 'react';
+import { act } from 'react-dom/test-utils';
 import { mount } from 'enzyme';
 import '../testSetup';
 import { run, useConcent, getState, register } from '../../src/index';